Update index; restructure layouts
