+++ database/static/frontpage.json
@@ -1,5 +1,6 @@
 {
   "headerHome": "Welcome!",
   "summaryBlockHome": "Our group is attempting to bring the same level of quality and content as technology meetups you would find in downtown Chicago. We hope to foster mentorship, networking, as well as skills and career growth.",
-  "quote": "Elevating other's only makes yourself stronger, to pull yourself up along.` - Unknown"
+  "quote": "When you help others rise, you help yourself grow.",
+  "quoteCite": "Matshona Dhliwayo"
 }
