+++ src/components/Layouts/LayoutWrapper.js
@@ -1,0 +1,42 @@
+/**
+ * Layout component that queries for data
+ * with Gatsby's useStaticQuery component
+ *
+ * See: https://www.gatsbyjs.org/docs/use-static-query/
+ */
+
+import React from 'react';
+import PropTypes from 'prop-types';
+import { useStaticQuery, graphql } from 'gatsby';
+
+import Header from '../Header';
+import Footer from '../Footer';
+import '../../assets/stylesheets/style.css';
+
+const LayoutWrapper = ({ children }) => {
+  const data = useStaticQuery(graphql`
+    query SiteTitleQuery {
+      site {
+        siteMetadata {
+          title
+        }
+      }
+    }
+  `);
+
+  return (
+    <>
+      <Header siteTitle={data.site.siteMetadata.title} />
+      <section className="flex-grow md:bg-grey-100 md:py-10">
+        {children}
+      </section>
+      <Footer />
+    </>
+  );
+};
+
+LayoutWrapper.propTypes = {
+  children: PropTypes.node.isRequired,
+};
+
+export default LayoutWrapper;
