+++ src/components/Layouts/index.js
@@ -1,0 +1,4 @@
+import LayoutWrapper from './LayoutWrapper';
+import DefaultLayout from './DefaultLayout';
+
+export { LayoutWrapper, DefaultLayout };
