+++ src/pages/404.js
@@ -1,15 +1,14 @@
-import React from "react";
+import React from 'react';
 
-import Layout from "../components/layout";
-import SEO from "../components/seo";
-
+import { DefaultLayout } from '../components/Layouts';
+import SEO from '../components/SEO';
 
 const NotFoundPage = () => (
-  <Layout>
+  <DefaultLayout>
     <SEO title="404: Not found" />
     <h1>NOT FOUND</h1>
     <p>You just hit a route that doesn&#39;t exist... the sadness.</p>
-  </Layout>
+  </DefaultLayout>
 );
 
 export default NotFoundPage;
