+++ src/pages/about-us.js
@@ -1,17 +1,19 @@
-import React from "react";
-import { Link } from "gatsby";
+import React from 'react';
+import { Link } from 'gatsby';
 
-import Layout from "../components/layout";
-import SEO from "../components/seo";
-
+import { DefaultLayout } from '../components/Layouts';
+import SEO from '../components/SEO';
 
 const AboutUs = () => (
-  <Layout>
+  <DefaultLayout>
     <SEO title="About Us" />
     <h1>About Us</h1>
-
+    {
+      // TODO: Add list of cities & villages (for SEO discovery)
+      // List available at http://www.lakecountyil.gov/BusinessDirectoryII.aspx?lngBusinessCategoryID=32
+    }
     <Link to="/">Go back to the homepage</Link>
-  </Layout>
+  </DefaultLayout>
 );
 
 export default AboutUs;
