+++ src/pages/index.js
@@ -1,32 +1,29 @@
-import React from "react";
+import React from 'react';
 
 import { useFetchJson } from '../hooks/useFetchJson';
-import Layout from "../components/layout";
-import Image from "../components/image";
-import SEO from "../components/seo";
-
+import { DefaultLayout } from '../components/Layouts';
+import SEO from '../components/SEO';
 
 const IndexPage = () => (
-
-  <Layout>
+  <DefaultLayout>
     <SEO title="Home" />
     <section>
       <h1>{useFetchJson(`headerHome`)}</h1>
       <p>{useFetchJson(`summaryBlockHome`)}</p>
       {
-      // ? We can change the quote however I think a good quote would be well placed here.
-      // I also suggest Better Comments extension if using VSCode because I use it a lot.
-      // Link: https://marketplace.visualstudio.com/items?itemName=aaron-bond.better-comments - Jacob
+        // ? We can change the quote however I think a good quote would be well placed here.
+        // I also suggest Better Comments extension if using VSCode because I use it a lot.
+        // Link: https://marketplace.visualstudio.com/items?itemName=aaron-bond.better-comments - Jacob
       }
-      <q style={{ fontFamily: `Roboto`, fontWeight: `bold` }}>{useFetchJson(`quote`)}</q>
+      <q className="text-center block">{useFetchJson(`quote`)}</q>
+      <cite className="text-center block">{useFetchJson(`quoteCite`)}</cite>
     </section>
-    <div style={{ maxWidth: `500px`, marginBottom: `1.45rem` }} alt="image icons of programming technologies">
-      <Image />
-    </div>
-    <div>PLACEHOLDER FOR MAP AND VENUE LOCATION INFO</div>
-    <div>PLACEHOLDER CONTACT INFO</div>
-    <div>PLACEHOLDER SUBMISSION FIELD FOR EMAIL/CONTACT US</div>
-  </Layout>
+    {
+      // TODO: Venue information and map
+      // TODO: Contact information
+      // TODO: Submission field for email / Contact Us
+    }
+  </DefaultLayout>
 );
 
 export default IndexPage;
